Add files via upload
Uploaded first.c and the Makefile for it. Pretty basic prompt. Finding which number is prime or not. Just used a loop that started at 2 and went up to half of the integer and mod'd it to see if there is any point where it is divisible and the % operation returns 0. If it returned 0 that means it is divisible by something other than 1 and itself.